apis/v1: Split Alertmanager types in their own file
